Remove Lines and finalizer from TextFileLineReader
As _lines is never updated, Lines always returns null.
This IS a breaking change but a reasonable one to do.

Also remove finalizer as TextFileLineReader doesn't hold any references to
unmanaged objects, which is what a finalizer is responsible for releasing.
Sure, it holds a reference to FileStream but it is managed and it's the one
holding references  to unmanged objects, and it has a destructor.