Skip Byte Order Marks at the start of text files and autodetect encoding
Automatically detect encoding and switch encoding based on Byte Order Marks
at the start of file.

The TextFileLineReader starts by reading the 4 first bytes, and looks for
UTF-8, UTF-16 (Big and Little endian) and UTF-32 (Big and Little endian).
If a BOM is found it's consumed and will not be sent to the user.
If the BOM is for an encoding different than the encoding specified by the user
in the constructor, a new BinaryReader with correct encoding is created, unless
the user has requested to not change encoding when creating the reader.

The detection behavior can be overriden by passing using another
IEncodingDetector. For example the NoEncodingDetector will ignore all bytes
in the beginning of a file.

One alternative considered but dismissed was to switch from BinaryReader to
a StreamReader, which automatically detects encoding, but instead of just
skipping the BOMs it removes them and set position to 0. So if the stream
contains a 3 bytes BOM and then an "a", BinaryReader would be at position 3
after reading the BOM and have "a" as next char to read, while StreamReader
woud be at position 0 and have "a" as next char. So switching reader would
mean that all stored positions in IStateStorage that were saved based on
BinaryReader would be wrong if used on StreamReader.

As MSpec has no support for datadriven tests xUnit is used for the tests.